Fix for prefetch failing when fk_id is 0
I have a table in which foreign key numbered 0 is a valid value.
Prefetch always sets None as `fk_id` and NoneAwaitable for the foreign model itself. It happens because `getattr` returns `0` an Python interprets it as `False` here.
This little change fixes that.

How correct do you think this is?